Added an ugly hack (usleep(1)) which calculating the LRU aging timestamp.
Desc. of Fix:
An ugly hack to prevent the same-time for different references
due to a usecond timer bug in non-QA code. The perfomance is still
reasonable;
i.e., under 7s for an input trace of 100k references.

QA code uses a lot of dprints, thus the time interval between
two consecutinve references of a same block is substantially
higher than in non-QA code.