Fixed incorrect mref while writing off a dirty block from L1 to L2.
Bug Description:
When a dirty tag is evicted from L1 by LRU (due to no space for the new
L1 miss), it has to be written back to next cache; L2, in our case.

While doing so, the incoming address was passed onto L2, which is incorrect.
The address of the dirty tag should be passed to L2.

Fix Description:
    * Added a new helper routine that converts a <tag, index> pair into
      a mem_ref_t structure. Please note that the ref_addr inside the
      mem_ref_t created by the helper routine will have it's block offset
      bits zeroed out.
    * Use the new helper to determine the addr of the dirty tag and pass
      it onto L2 while evicting it.